feat: status skeleton
This will prevent discord embend to fail if request takes longer then timeout for discord embend to receive an interaction
